+++ crates/rust-client/src/note/mod.rs
@@ -90,7 +90,9 @@ pub use miden_objects::{
     },
 };
 pub use note_screener::{NoteConsumability, NoteRelevance, NoteScreener, NoteScreenerError};
-pub use note_update_tracker::NoteUpdateTracker;
+pub use note_update_tracker::{
+    InputNoteUpdate, NoteUpdateTracker, NoteUpdateType, OutputNoteUpdate,
+};
 
 /// Note retrieval methods.
 impl Client {